Refactor node type

+++ src/com/ido/pipeline/app/AndroidAppPipeline.groovy
@@ -12,14 +12,18 @@ class AndroidAppPipeline extends JdkPipeline {
 
     AndroidAppPipeline(Object steps) {
         super(steps)
+        this.useK8sAgent = true
+        this.nodeName = "android"
     }
 
     @Override
     Map runPipeline(Map config) {
-        String builder = steps.libraryResource(resource: 'pod-template/android-builder.yaml', encoding: 'UTF-8')
-        builder = builder.replaceAll('<builderImage>', config.java.builderImage)
-                .replaceAll('<androidCmdLineToolsUrl>', config._system.android.cmdLineToolsUrl)
-        config.podTemplate = builder
+        if (config.useK8sAgent == null || config.useK8sAgent) {
+            String builder = steps.libraryResource(resource: 'pod-template/android-builder.yaml', encoding: 'UTF-8')
+            builder = builder.replaceAll('<builderImage>', config.java.builderImage)
+                    .replaceAll('<androidCmdLineToolsUrl>', config._system.android.cmdLineToolsUrl)
+            config.podTemplate = builder
+        }
 
         return super.runPipeline(config)
     }
@@ -42,36 +46,37 @@ class AndroidAppPipeline extends JdkPipeline {
         if (config._system.android.proxy.enable) {
             proxy = "--no_https --proxy=http --proxy_host=${config._system.android.proxy.host} --proxy_port=${config._system.android.proxy.port}"
         }
-        steps.container('builder') {
-            List<String> sdkPackagesInstalledOutput = steps.sh(returnStdout: true, encoding: "UTF-8",
-                    script: """${config.debugSh}
-\${ANDROID_HOME}/cmdline-tools/latest/bin/sdkmanager --list_installed | tail -n +4 | awk '{print \$1}'
-""")
+        List<String> sdkPackagesInstalledOutput = execOnAgent('builder', {
+            steps.sh(returnStdout: true, encoding: "UTF-8", script: """${config.debugSh}
+                \${ANDROID_HOME}/cmdline-tools/latest/bin/sdkmanager --list_installed | tail -n +4 | awk '{print \$1}'
+            """)
                     .trim()
                     .split(System.lineSeparator()) as List<String>
+        })
 
-            Boolean start = false
-            def sdkPackagesInstalled = []
-            for (int i=0; i < sdkPackagesInstalledOutput.size(); i++) {
-                if (start) {
-                    sdkPackagesInstalled.add(sdkPackagesInstalledOutput[i])
-                }
-                if (sdkPackagesInstalledOutput[i].matches("^\\s*----.*")) {
-                    start = true
-                }
+        Boolean start = false
+        def sdkPackagesInstalled = []
+        for (int i = 0; i < sdkPackagesInstalledOutput.size(); i++) {
+            if (start) {
+                sdkPackagesInstalled.add(sdkPackagesInstalledOutput[i])
             }
-
-            steps.echo "sdkPackagesInstalled: ${sdkPackagesInstalled}"
-
-            List<String> sdkPackagesToBeInstalled = (config.android.sdkPackagesRequired as List<String>)
-                    .findAll {!sdkPackagesInstalled.contains(it)}
-
-            String sdkPackages = ""
-            sdkPackagesToBeInstalled.each {
-                sdkPackages = sdkPackages + "'$it' "
+            if (sdkPackagesInstalledOutput[i].matches("^\\s*----.*")) {
+                start = true
             }
-            steps.echo "sdkPackages: ${sdkPackages}"
+        }
 
+        steps.echo "sdkPackagesInstalled: ${sdkPackagesInstalled}"
+
+        List<String> sdkPackagesToBeInstalled = (config.android.sdkPackagesRequired as List<String>)
+                .findAll { !sdkPackagesInstalled.contains(it) }
+
+        String sdkPackages = ""
+        sdkPackagesToBeInstalled.each {
+            sdkPackages = sdkPackages + "'$it' "
+        }
+        steps.echo "sdkPackages: ${sdkPackages}"
+
+        execOnAgent('builder', {
             steps.sh """${config.debugSh}
                 if [ ! -e \${ANDROID_HOME}/licenses-agreed ]; then
                     yes | \${ANDROID_HOME}/cmdline-tools/latest/bin/sdkmanager ${proxy} --licenses
@@ -81,23 +86,23 @@ class AndroidAppPipeline extends JdkPipeline {
                     yes | \${ANDROID_HOME}/cmdline-tools/latest/bin/sdkmanager ${proxy} --install $sdkPackages
                 fi
             """
-        }
+        })
     }
 
     @Override
     def build() {
         String newFileName = fileArchiver.getFileName()
-        steps.container('builder') {
-            String updateDependenciesArgs = ""
-            if (config.java.forceUpdateDependencies) {
-                updateDependenciesArgs = "--refresh-dependencies"
-            }
+        String updateDependenciesArgs = ""
+        if (config.java.forceUpdateDependencies) {
+            updateDependenciesArgs = "--refresh-dependencies"
+        }
 
-            String initScriptArgs = ""
-            if (config.java.defaultGradleInitScript) {
-                initScriptArgs = "-I \"${steps.env.WORKSPACE}/${config.srcRootPath}/default-gradle-init.gradle\""
-            }
+        String initScriptArgs = ""
+        if (config.java.defaultGradleInitScript) {
+            initScriptArgs = "-I \"${steps.WORKSPACE}/${config.srcRootPath}/default-gradle-init.gradle\""
+        }
 
+        execOnAgent('builder', {
             steps.sh """${config.debugSh}
                 cd "${config.srcRootPath}"
                 mkdir -p ido-cluster/outputs/files
@@ -112,7 +117,7 @@ class AndroidAppPipeline extends JdkPipeline {
                         -Dfile.encoding=UTF-8 \
                         "-Dorg.gradle.jvmargs=-Xmx2048m -XX:+HeapDumpOnOutOfMemoryError -Dfile.encoding=UTF-8" \
                         -p ${config.java.moduleName}
-                    numFound=\$(find ./ -type f -name "*.apk" -print | wc -l)
+                    numFound=\$(find ./ -type f -name "*.apk" -not -path "./ido-cluster/*" -print | wc -l)
                     if [ \$numFound -lt 1 ]; then
                         echo "Can't find debug version apk file!"
                         exit 1
@@ -121,7 +126,7 @@ class AndroidAppPipeline extends JdkPipeline {
                         echo "Find multiple debug version apk files!"
                         exit 1
                     fi
-                    find ./ -type f -name "*.apk" \
+                    find ./ -type f -name "*.apk" -not -path "./ido-cluster/*" \
                         -exec mv "{}" "ido-cluster/outputs/files/${newFileName}-debug.apk" \\;
                 fi
                 if [ "$config.android.buildRelease" = "true" ]; then
@@ -133,7 +138,7 @@ class AndroidAppPipeline extends JdkPipeline {
                         -Dfile.encoding=UTF-8 \
                         "-Dorg.gradle.jvmargs=-Xmx2048m -XX:+HeapDumpOnOutOfMemoryError -Dfile.encoding=UTF-8" \
                         -p ${config.java.moduleName}
-                    numFound=\$(find ./ -type f -name "*.apk" -print | wc -l)
+                    numFound=\$(find ./ -type f -name "*.apk" -not -path "./ido-cluster/*" -print | wc -l)
                     if [ \$numFound -lt 1 ]; then
                         echo "Can't find release version apk file!"
                         exit 1
@@ -142,25 +147,30 @@ class AndroidAppPipeline extends JdkPipeline {
                         echo "Find multiple release version apk files!"
                         exit 1
                     fi
-                    find ./ -type f -name "*.apk" \
+                    find ./ -type f -name "*.apk" -not -path "./ido-cluster/*" \
                         -exec mv "{}" "ido-cluster/outputs/files/${newFileName}-release.apk" \\;
                 fi
             """
-        }
+        })
     }
 
     @Override
     def archive() {
         String newFileName = fileArchiver.getFileName()
-        steps.container('uploader') {
-            String uploadUrl = "${config.fileServer.uploadUrl}${config.fileServer.uploadRootPath}${config.productName}/" +
-                    config.branch + "/android"
-            String debugDownloadUrl = "${config.fileServer.downloadUrl}${config.fileServer.uploadRootPath}/${config.productName}/" +
-                    config.branch + "/android/files/${newFileName}-debug.apk"
-            String releaseDownloadUrl = "${config.fileServer.downloadUrl}${config.fileServer.uploadRootPath}/${config.productName}/" +
-                    config.branch + "/android/files/${newFileName}-release.apk"
+        String uploadUrl = "${config.fileServer.uploadUrl}${config.fileServer.uploadRootPath}${config.productName}/" +
+                config.branch + "/android"
 
+        if (!config.useK8sAgent) {
+            String text = steps.libraryResource('tools/generate_qrcode.py')
+            steps.writeFile(file: "generate_qrcode.py", text: text, encoding: "UTF-8")
+        }
 
+        String debugDownloadUrl = "${config.fileServer.downloadUrl}${config.fileServer.uploadRootPath}/${config.productName}/" +
+                config.branch + "/android/files/${newFileName}-debug.apk"
+        String releaseDownloadUrl = "${config.fileServer.downloadUrl}${config.fileServer.uploadRootPath}/${config.productName}/" +
+                config.branch + "/android/files/${newFileName}-release.apk"
+
+        execOnAgent('uploader', {
             steps.sh """${config.debugSh}
                 cd "${config.srcRootPath}/ido-cluster/outputs"
                 touch ${newFileName}.html
@@ -169,34 +179,44 @@ class AndroidAppPipeline extends JdkPipeline {
                 
                 if [ "${config.android.buildDebug}" = "true" ]; then
                     echo "debugDownloadUrl: ${debugDownloadUrl}"
-                    qrencode --output qrcode.png "${debugDownloadUrl}"
-                    if [ ! -f qrcode.png ]; then
-                        echo QR code is not generated!
-                        exit 1
+
+                    if [ "${config.useK8sAgent}" = "true" ]; then
+                        qrencode --output qrcode.png "${debugDownloadUrl}"
+                        if [ ! -f qrcode.png ]; then
+                            echo QR code is not generated!
+                            exit 1
+                        fi
+                        debugQrcode="\$(cat qrcode.png | base64 )"
+                        rm -f qrcode.png
+                    else
+                        debugQrcode=\$(python3 ${steps.WORKSPACE}/generate_qrcode.py "${debugDownloadUrl}")
                     fi
-                    debugQrcode="\$(cat qrcode.png | base64 )"
-                    rm -f qrcode.png
-                    
+
                     echo "<hr><p><a href='files/${newFileName}-debug.apk'>${newFileName}-debug.apk</a>" >> ${newFileName}.html
                     echo "<p><img src='data:image/png;base64,\${debugQrcode}'/>" >> ${newFileName}.html
                 fi
                 
                 if [ "${config.android.buildRelease}" = "true" ]; then
                     echo "releaseDownloadUrl: ${releaseDownloadUrl}"
-                    qrencode --output qrcode.png "${releaseDownloadUrl}"
-                    if [ ! -f qrcode.png ]; then
-                        echo QR code is not generated!
-                        exit 1
+
+                    if [ "${config.useK8sAgent}" = "true" ]; then
+                        qrencode --output qrcode.png "${releaseDownloadUrl}"
+                        if [ ! -f qrcode.png ]; then
+                            echo QR code is not generated!
+                            exit 1
+                        fi
+                        releaseQrcode="\$(cat qrcode.png | base64 )"
+                        rm -f qrcode.png
+                    else
+                        releaseQrcode=\$(python3 ${steps.WORKSPACE}/generate_qrcode.py "${releaseDownloadUrl}")
                     fi
-                    releaseQrcode="\$(cat qrcode.png | base64 )"
-                    rm -f qrcode.png
                     
                     echo "<hr><p><a href='files/${newFileName}-release.apk'>${newFileName}-release.apk</a>" >> ${newFileName}.html
                     echo "<p><img src='data:image/png;base64,\${releaseQrcode}'/>" >> ${newFileName}.html
                 fi
             """
+        })
 
-            fileArchiver.upload(uploadUrl, "${config.srcRootPath}/ido-cluster/outputs")
-        }
+        fileArchiver.upload(uploadUrl, "${config.srcRootPath}/ido-cluster/outputs")
     }
 }
